Make UpsTokenPair own class at dedicated folder
